Move Android definitions to own source file
